This is version 0.4 shipped to CRAN

+++ reverse_dependencies/foo.R
@@ -1,8 +1,10 @@
 # clean
 foo <- list.files()
-bar <- grep("^foo\\.R", foo, invert = TRUE, value = TRUE)
-baz <- paste("rm -r", paste(bar, collapse = " "))
-system(baz)
+bar <- setdiff(foo, c("foo.R", "foo.Rout"))
+if (length(bar) > 0) {
+    baz <- paste("rm -r", paste(bar, collapse = " "))
+    system(baz)
+}
 list.files()
 # my package
 system("R CMD build ../package/polyapost")
